Add test coverage for omise-block-credit-card.php
