+++ tests/unit/includes/blocks/gateways/omise-block-credit-card-test.php
@@ -6,7 +6,7 @@ use Brain\Monkey;
 
 require_once __DIR__ . '/traits/mock-gateways.php';
 
-class Omise_Block_Apm_Test extends TestCase
+class Omise_Block_Credit_Card_Test extends TestCase
 {
     // Adds Mockery expectations to the PHPUnit assertions count.
     use MockeryPHPUnitIntegration, MockPaymentGateways;
@@ -18,8 +18,8 @@ class Omise_Block_Apm_Test extends TestCase
     {
         parent::setUp();
         $this->mockWcGateways();
-        require_once __DIR__ . '/../../../../../includes/blocks/gateways/abstract-omise-block-apm.php';
-        $this->obj = new class extends Omise_Block_Apm {};
+        require_once __DIR__ . '/../../../../../includes/blocks/gateways/omise-block-credit-card.php';
+        $this->obj = new Omise_Block_Credit_Card;
     }
 
     /**
@@ -29,13 +29,9 @@ class Omise_Block_Apm_Test extends TestCase
     {
         Monkey\Functions\expect('get_option')->andReturn(null);
 
-        $reflection = new \ReflectionClass($this->obj);
-        $name_property = $reflection->getProperty('name');
-        $name_property->setAccessible(true);
-        $name_property->setValue($this->obj, 'omise_promptpay');
-
         $this->obj->initialize();
 
+        $reflection = new \ReflectionClass($this->obj);
         $gateway_property = $reflection->getProperty('gateway');
         $gateway_property->setAccessible(true);
         $gateway_val = $gateway_property->getValue($this->obj);
@@ -48,12 +44,6 @@ class Omise_Block_Apm_Test extends TestCase
      */ 
     public function is_active()
     {
-        // Calling initialize() to set $gateway value
-        $reflection = new \ReflectionClass($this->obj);
-        $name_property = $reflection->getProperty('name');
-        $name_property->setAccessible(true);
-        $name_property->setValue($this->obj, 'omise_promptpay');
-
         $this->obj->initialize();
 
         $is_active = $this->obj->is_active();
@@ -61,7 +51,6 @@ class Omise_Block_Apm_Test extends TestCase
     }
 
     /**
-     * @test
      */
     public function get_payment_method_data()
     {
@@ -69,7 +58,7 @@ class Omise_Block_Apm_Test extends TestCase
         $reflection = new \ReflectionClass($this->obj);
         $name_property = $reflection->getProperty('name');
         $name_property->setAccessible(true);
-        $name_property->setValue($this->obj, 'omise_promptpay');
+        $name_property->setValue($this->obj, 'abc');
 
         $this->obj->initialize();
 
@@ -79,7 +68,7 @@ class Omise_Block_Apm_Test extends TestCase
         $this->assertArrayHasKey('description', $data);
         $this->assertArrayHasKey('supports', $data);
         $this->assertEquals('array', gettype($data['supports']));
-        $this->assertEquals('omise_promptpay', $data['name']);
+        $this->assertEquals('abc', $data['name']);
     }
 
     /**
@@ -90,10 +79,13 @@ class Omise_Block_Apm_Test extends TestCase
         Monkey\Functions\expect('wp_script_is')->andReturn(false);
         Monkey\Functions\expect('wp_register_script');
         Monkey\Functions\expect('plugin_dir_url');
-        Monkey\Functions\expect('wp_enqueue_script');
+        Monkey\Functions\expect('is_checkout')->andReturn(true);
+        // Monkey\Functions\expect('is_user_logged_in')->andReturn(false);
+
+        $this->obj->initialize();
 
         $result = $this->obj->get_payment_method_script_handles();
 
-        $this->assertEquals([ 'wc-omise-one-click-apms-payments-blocks' ], $result);
+        $this->assertEquals([ 'omise-payments-blocks' ], $result);
     }
 }
